+++ apps/web/config/routes.rb
@@ -4,3 +4,4 @@ resources :errors,         only: [:index, :show], controller: "external_errors"
 resources :reservations,   only: [:index]
 resources :suppliers,      only: [:show]
 resources :sync_processes, only: [:index]
+